init(index): add prototype version to git
Add the initial prototype matrix decoder to the git repository.  This
was originally developed on rather an ad-hoc basis and so didn't
particularly warrant a git repository for an ephermal piece of code.
However, it appears this may be published and so it made sense to add it
to a versioned repository!